% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAge.R
\name{calcAge}
\alias{calcAge}
\title{Calculate 14C age}
\usage{
calcAge(logdata, edit_C14window = FALSE)
}
\arguments{
\item{logdata}{A list generated with \code{readRawdata}.}

\item{edit_C14window}{A logical scalar.
If TRUE, the editor is displayed and the C14 window can be edited.}
}
\description{
\code{calcAge} returns a list which includes 14C age and other values used for calculation
}
