% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpecSum.R
\name{plotSpecSum}
\alias{plotSpecSum}
\title{Plot 14C Spectrum}
\usage{
plotSpecSum(logdata)
}
\arguments{
\item{logdata}{A list generated with \code{readRawdata}.}
}
\description{
\code{plotSpecSum} returns html plot of 14C spectrum.
A plot for the number of measurements is output.
}
