% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRawdata.R
\name{readRawdata}
\alias{readRawdata}
\title{Read Raw Logfiles and Convert to List}
\usage{
readRawdata(path, index_path = NULL)
}
\arguments{
\item{path}{A string that indicates the path to the log file folder.}

\item{index_path}{A string that indicates the path to the index csv file.
If not, the index file is assumed to be in the folder indicated by @param path.}
}
\description{
\code{readRawdata} returns a list which includes all data in logfiles.
}
